Unify bulk run/processing scripts
