+++ experiments/process_bulk_evict_latency.py
@@ -5,6 +5,7 @@ import matplotlib.pyplot as plt
 import process_utility as u
 from run_bulk_evict_latency import functions, base_window_sizes, bulk_sizes
 import math
+import process_bulk_data_latency as pbdl
 
 aggregators = [
     "daba_lite",
@@ -32,7 +33,7 @@ def make_bulk_evict_latency_plots():
                     data[agg] = u.read_latency_data('results/' + file_preamble + '_' + agg + '_' + exp_name + '.txt')
 
                 print(exp_name)
-                u.make_violin_graph(data, aggregators, exp_name, exp_title, file_preamble)
+                pbdl.make_violin_graph(data, aggregators, exp_name, exp_title, file_preamble, f)
             
 def main():
     u.create_figures_dir()
